requirements.txt
opencv-python==4.8.1.78
mlx==0.8.0
mlx-lm==0.8.0
torch==2.1.0
torchvision==0.16.0
flask==2.3.3
pillow==10.0.1
numpy==1.24.3
python-dateutil==2.8.2
jinja2==3.1.2
werkzeug==2.3.7
transformers==4.35.0